ENH: Ability to specify ROI binary mask to confine volume rendering to
One may optionally, set a mask input. This mask may be a binary mask
or a label map. This must be specified via SetMaskType.

If the mask is a binary mask, the volume rendering is confined to regions
within the binary mask. The binary mask is assumed to have a datatype of
UCHAR and values of 255 (inside) and 0 (outside).

This is only implemented for the GPU volume mapper. And for the Composite
blend case. This will be expanded for the other blend cases in the next
commit.