Update second tecplot reader baseline
Commit d7fd23aa (Added second correct image for tecplot, 2010-07-22)
added a second baseline image for TestTecplotReader that accounts for
OpenGL implementations that flip the rasterization direction of the
color map.  Later commit f3bbe784 (Updating the base line image for
tecplot reader, 2010-09-08) updated the original baseline image for a
change in the test.  This commit updates the second baseline image for
the new version of the test.